feat(router): Change route of sensor and user to new formats
- /accounts/[username] becomes /[username]
- /sensors/[id] becomes /[username]/[id]

The redirects are taken care of using NextJS' redirect feature. (These
are not reflected in the sitemap.xml, only new routes will be present
there)

Certain words that can conflict with current, previous or future routes
are now not allowed as username while creating an account to prevent
unreachable routes.